+++ docs/source/tutorial/atomic_actions.rst
@@ -101,8 +101,9 @@ Focused examples live under ``scripts/tutorials/atomic_action``:
 * ``move_held_object.py``
 * ``place.py``
 * ``assemble.py``
-* ``press_button.py``
-* ``turn_knob.py``
+* ``press.py``
+* ``slide.py``
+* ``twist.py``
 * ``coordinated_pickment.py``
 * ``coordinated_placement.py``
 * ``hand_over.py``
