+++ embodichain/lab/sim/atomic_actions/__init__.py
@@ -33,9 +33,9 @@ from .affordance import (
     AntipodalAffordance,
     AssembleAffordance,
     InteractionPoints,
-    PressButtonAffordance,
-    PullPushAffordance,
-    TurnAffordance,
+    PressAffordance,
+    SlideAffordance,
+    TwistAffordance,
 )
 from .bindings import ActionBinding, ResolvedActionBinding, ResolvedControlPart
 from .control import (
@@ -109,15 +109,15 @@ from .primitives import (
     Place,
     PlaceGoal,
     PlaceOptions,
-    PressButton,
-    PressButtonGoal,
-    PressButtonOptions,
-    PullPushArticulatedPart,
-    PullPushArticulatedPartGoal,
-    PullPushArticulatedPartOptions,
-    TurnKnob,
-    TurnKnobGoal,
-    TurnKnobOptions,
+    Press,
+    PressGoal,
+    PressOptions,
+    Slide,
+    SlideGoal,
+    SlideOptions,
+    Twist,
+    TwistGoal,
+    TwistOptions,
 )
 from .runner import (
     CommandAcknowledgement,
@@ -231,17 +231,17 @@ __all__ = [
     "PlannerDiagnostics",
     "PlanningContext",
     "PoseGoalValue",
-    "PressButton",
-    "PressButtonAffordance",
-    "PressButtonGoal",
-    "PressButtonOptions",
-    "PullPushAffordance",
-    "PullPushArticulatedPart",
-    "PullPushArticulatedPartGoal",
-    "PullPushArticulatedPartOptions",
-    "TurnKnob",
-    "TurnKnobGoal",
-    "TurnKnobOptions",
+    "Press",
+    "PressAffordance",
+    "PressGoal",
+    "PressOptions",
+    "SlideAffordance",
+    "Slide",
+    "SlideGoal",
+    "SlideOptions",
+    "Twist",
+    "TwistGoal",
+    "TwistOptions",
     "RecoveryPolicy",
     "RigidObjectSceneProvider",
     "RigidObjectSceneProviderCfg",
@@ -264,6 +264,6 @@ __all__ = [
     "SimulationExecutionAdapter",
     "TaskState",
     "TimedTrajectory",
-    "TurnAffordance",
+    "TwistAffordance",
     "TrajectorySegment",
 ]
