+++ embodichain/lab/sim/atomic_actions/primitives/twist.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-"""PressButton atomic action implementation."""
+"""Twist atomic action implementation."""
 
 from __future__ import annotations
 
@@ -24,10 +24,16 @@ from typing import ClassVar
 
 import torch
 
-from embodichain.utils.math import get_relative_rotation
-from embodichain.lab.sim.atomic_actions.affordance import PressButtonAffordance
+from embodichain.utils.math import (
+    axis_angle_to_rotation_matrix,
+    pose_inv,
+    get_relative_rotation,
+)
+
+from embodichain.lab.sim.atomic_actions.affordance import TwistAffordance
 from embodichain.lab.sim.atomic_actions.control import (
     GRASP_COMMAND,
+    OPEN_COMMAND,
     JointPositionCommand,
 )
 from embodichain.lab.sim.atomic_actions.core import AtomicAction, ObjectSemantics
@@ -58,56 +64,45 @@ from embodichain.lab.sim.atomic_actions.trajectory_ops import (
 
 
 @dataclass(frozen=True, slots=True, eq=False)
-class PressButtonGoal(ObjectActionGoal):
-    """Articulation-link or rigid button described by a press affordance."""
+class TwistGoal(ObjectActionGoal):
+    """Target object described by a twist affordance."""
 
-    goal_kind: ClassVar[str] = "press_button"
+    goal_kind: ClassVar[str] = "twist"
 
 
 @dataclass(frozen=True, slots=True, eq=False)
-class PressButtonOptions(ActionOptions):
-    """Per-invocation button-pressing behavior."""
+class TwistOptions(ActionOptions):
+    """Per-invocation twisting behavior."""
 
     hand_interp_steps: int = 5
-    """Number of waypoints used to close the hand."""
+    """Number of waypoints used for each close/open hand segment."""
 
-    approach_distance: float = 0.1
-    """Distance from the press position opposite the press direction."""
+    twist_waypoint_count: int = 8
+    """Number of Cartesian keyframes along the target's circular twist arc."""
 
-    press_distance: float = 0.05
-    """Distance traveled into the button along its press axis."""
+    pre_grasp_distance: float = 0.1
+    """Distance from the grasp pose along its negative z-axis."""
 
-    press_position: tuple[float, float, float] | None = None
-    """Optional local-frame position overriding the affordance press position."""
+    twist_angle: float = math.pi / 4
+    """Requested twist rotation in radians."""
 
     def __post_init__(self) -> None:
         if self.hand_interp_steps < 1:
             raise ValueError("hand_interp_steps must be at least 1.")
-        if not math.isfinite(self.approach_distance):
-            raise ValueError("approach_distance must be finite.")
-        if self.approach_distance < 0.0:
-            raise ValueError("approach_distance must be non-negative.")
-        if not math.isfinite(self.press_distance):
-            raise ValueError("press_distance must be finite.")
-        if self.press_distance <= 0.0:
-            raise ValueError("press_distance must be positive.")
-        if self.press_position is not None:
-            position = torch.as_tensor(self.press_position, dtype=torch.float32)
-            if position.shape != (3,) or not torch.isfinite(position).all():
-                raise ValueError("press_position must be a finite (x, y, z) tuple.")
-            object.__setattr__(
-                self,
-                "press_position",
-                tuple(float(component) for component in position),
-            )
+        if self.twist_waypoint_count < 1:
+            raise ValueError("twist_waypoint_count must be at least 1.")
+        if self.pre_grasp_distance < 0.0:
+            raise ValueError("pre_grasp_distance must be non-negative.")
+        if not math.isfinite(self.twist_angle):
+            raise ValueError("twist_angle must be finite.")
 
 
-class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
-    """Close the gripper, approach and press a button, then retract."""
+class Twist(AtomicAction[TwistGoal, TwistOptions]):
+    """Approach, grasp, twist, release, and retract from a target."""
 
-    skill_id: ClassVar[str] = "press_button"
-    GoalType: ClassVar[type] = PressButtonGoal
-    OptionsType: ClassVar[type] = PressButtonOptions
+    skill_id: ClassVar[str] = "twist"
+    GoalType: ClassVar[type] = TwistGoal
+    OptionsType: ClassVar[type] = TwistOptions
     manipulator_roles: ClassVar[tuple[str, ...]] = ("primary",)
     end_effector_roles: ClassVar[tuple[str, ...]] = ("primary",)
     binding_contract: ClassVar[SkillBindingContract] = SkillBindingContract(
@@ -128,7 +123,10 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
                     SkillEndpointRequirement(
                         endpoint_id="grasp",
                         capabilities=frozenset({GRASP_CAPABILITY}),
-                        required_commands={GRASP_COMMAND: JointPositionCommand},
+                        required_commands={
+                            OPEN_COMMAND: JointPositionCommand,
+                            GRASP_COMMAND: JointPositionCommand,
+                        },
                         route=ActionBindingRoute("end_effector", "primary"),
                     ),
                 ),
@@ -137,7 +135,7 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
         ),
     )
 
-    def __init__(self, default_options: PressButtonOptions | None = None) -> None:
+    def __init__(self, default_options: TwistOptions | None = None) -> None:
         super().__init__(default_options)
 
     def _on_bind(self) -> None:
@@ -164,19 +162,24 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
 
     def _plan(
         self,
-        request: ResolvedActionRequest[PressButtonGoal, PressButtonOptions],
+        request: ResolvedActionRequest[TwistGoal, TwistOptions],
         context: PlanningContext,
     ) -> ActionPlan:
-        """Plan close, approach, press, and retract without stepping simulation."""
+        """Plan all six twisting segments without stepping simulation."""
         target = self.require_goal(request)
-        affordance = self._require_press_button_affordance(target.semantics)
+        affordance = self._require_twist_affordance(target.semantics)
         options = request.skill_options
         manipulator = request.binding.manipulator()
         end_effector = request.binding.end_effector()
         arm_joint_ids = list(manipulator.joint_ids)
         hand_joint_ids = list(end_effector.joint_ids)
         start_arm_qpos = arm_qpos_from_state(context, arm_joint_ids)
-        start_hand_qpos = context.last_qpos[:, hand_joint_ids]
+        hand_open_qpos = end_effector.joint_positions(
+            OPEN_COMMAND,
+            num_envs=context.batch_size,
+            device=self.device,
+            dtype=context.robot.qpos.dtype,
+        )
         hand_grasp_qpos = end_effector.joint_positions(
             GRASP_COMMAND,
             num_envs=context.batch_size,
@@ -189,60 +192,83 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
         )
         if link_pose.shape != (self.num_envs, 4, 4):
             raise ValueError(
-                "Button target pose must have shape "
+                "Twist target pose must have shape "
                 f"({self.num_envs}, 4, 4), got {tuple(link_pose.shape)}."
             )
-        contact_xpos = affordance.get_press_pose(
-            link_pose,
-            press_position=options.press_position,
-        ).to(device=self.device, dtype=torch.float32)
-        contact_xpos = self._find_symmetric_nearest_xpos(
-            contact_xpos,
+        grasp_xpos = affordance.get_grasp_pose(link_pose).to(
+            device=self.device, dtype=torch.float32
+        )
+        grasp_xpos = self._find_symmetric_nearest_xpos(
+            grasp_xpos,
             reference_xpos=self.robot.compute_fk(
                 qpos=start_arm_qpos, name=manipulator.name, to_matrix=True
             ),
         )
-        approach_xpos = translate_pose_world(
-            contact_xpos,
-            -contact_xpos[:, :3, 2] * options.approach_distance,
+        pre_grasp_xpos = translate_pose_world(
+            grasp_xpos,
+            -grasp_xpos[:, :3, 2] * options.pre_grasp_distance,
         )
-        pressed_xpos = translate_pose_world(
-            contact_xpos,
-            contact_xpos[:, :3, 2] * options.press_distance,
+        twist_xpos = self._twisted_grasp_poses(
+            link_pose,
+            grasp_xpos,
+            affordance.twist_axis,
+            options.twist_angle,
+            options.twist_waypoint_count,
         )
-        n_approach, n_press, n_retract = self._motion_segment_lengths(
+
+        n_approach, n_reach, n_twist, n_retract = self._motion_segment_lengths(
             request.motion_policy.sample_count,
             options.hand_interp_steps,
         )
-        hand_close = interpolate_hand_qpos(
-            start_hand_qpos,
-            hand_grasp_qpos,
-            n_waypoints=options.hand_interp_steps,
-        )
+
         approach_success, approach_arm = self._plan_pose_segment(
-            approach_xpos,
+            pre_grasp_xpos,
             start_arm_qpos,
             manipulator.name,
             request,
             n_approach,
         )
-        press_success, press_arm = self._plan_pose_segment(
-            pressed_xpos,
+        reach_success, reach_arm = self._plan_pose_segment(
+            grasp_xpos,
             approach_arm[:, -1],
             manipulator.name,
             request,
-            n_press,
+            n_reach,
+        )
+        twist_success, twist_arm = self._plan_pose_segment(
+            twist_xpos,
+            reach_arm[:, -1],
+            manipulator.name,
+            request,
+            n_twist,
         )
         retract_success, retract_arm = self._plan_pose_segment(
-            approach_xpos,
-            press_arm[:, -1],
+            pre_grasp_xpos,
+            twist_arm[:, -1],
             manipulator.name,
             request,
             n_retract,
         )
-        success = approach_success & press_success & retract_success
+        success = approach_success & reach_success & twist_success & retract_success
 
-        parts = (hand_close, approach_arm, press_arm, retract_arm)
+        hand_close = interpolate_hand_qpos(
+            hand_open_qpos,
+            hand_grasp_qpos,
+            n_waypoints=options.hand_interp_steps,
+        )
+        hand_open = interpolate_hand_qpos(
+            hand_grasp_qpos,
+            hand_open_qpos,
+            n_waypoints=options.hand_interp_steps,
+        )
+        parts = (
+            approach_arm,
+            reach_arm,
+            hand_close,
+            twist_arm,
+            hand_open,
+            retract_arm,
+        )
         lengths = tuple(part.shape[1] for part in parts)
         full = torch.empty(
             (self.num_envs, sum(lengths), self.robot_dof),
@@ -251,17 +277,32 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
         )
         full[:] = context.last_qpos.unsqueeze(1)
         offset = 0
-
-        stop = offset + hand_close.shape[1]
-        full[:, offset:stop, arm_joint_ids] = start_arm_qpos.unsqueeze(1)
-        full[:, offset:stop, hand_joint_ids] = hand_close
-        offset = stop
-
-        for arm in (approach_arm, press_arm, retract_arm):
+        arm_parts = (approach_arm, reach_arm, twist_arm, retract_arm)
+        arm_hands = (
+            hand_open_qpos,
+            hand_open_qpos,
+            hand_grasp_qpos,
+            hand_open_qpos,
+        )
+        for arm, hand in zip(arm_parts[:2], arm_hands[:2]):
             stop = offset + arm.shape[1]
             full[:, offset:stop, arm_joint_ids] = arm
-            full[:, offset:stop, hand_joint_ids] = hand_grasp_qpos.unsqueeze(1)
+            full[:, offset:stop, hand_joint_ids] = hand.unsqueeze(1)
             offset = stop
+        stop = offset + hand_close.shape[1]
+        full[:, offset:stop, arm_joint_ids] = reach_arm[:, -1].unsqueeze(1)
+        full[:, offset:stop, hand_joint_ids] = hand_close
+        offset = stop
+        stop = offset + twist_arm.shape[1]
+        full[:, offset:stop, arm_joint_ids] = twist_arm
+        full[:, offset:stop, hand_joint_ids] = hand_grasp_qpos.unsqueeze(1)
+        offset = stop
+        stop = offset + hand_open.shape[1]
+        full[:, offset:stop, arm_joint_ids] = twist_arm[:, -1].unsqueeze(1)
+        full[:, offset:stop, hand_joint_ids] = hand_open
+        offset = stop
+        full[:, offset:, arm_joint_ids] = retract_arm
+        full[:, offset:, hand_joint_ids] = hand_open_qpos.unsqueeze(1)
 
         return self.build_plan(
             request,
@@ -270,43 +311,45 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
             trajectory=full,
             expected_effects=StateDelta(),
             segment_lengths={
-                "close": lengths[0],
-                "approach": lengths[1],
-                "press": lengths[2],
-                "retract": lengths[3],
+                "approach": lengths[0],
+                "reach": lengths[1],
+                "close": lengths[2],
+                "twist": lengths[3],
+                "open": lengths[4],
+                "retract": lengths[5],
             },
         )
 
     @staticmethod
-    def _require_press_button_affordance(
+    def _require_twist_affordance(
         semantics: ObjectSemantics,
-    ) -> PressButtonAffordance:
+    ) -> TwistAffordance:
         affordance = semantics.affordance
-        if not isinstance(affordance, PressButtonAffordance):
-            raise ValueError("PressButton requires a PressButtonAffordance.")
+        if not isinstance(affordance, TwistAffordance):
+            raise ValueError("Twist requires a TwistAffordance.")
         return affordance
 
     @staticmethod
     def _motion_segment_lengths(
         sample_count: int,
         hand_interp_steps: int,
-    ) -> tuple[int, int, int]:
-        motion_count = sample_count - hand_interp_steps
-        if motion_count < 6:
+    ) -> tuple[int, int, int, int]:
+        motion_count = sample_count - 2 * hand_interp_steps
+        if motion_count < 8:
             raise ValueError(
-                "Not enough waypoints for PressButton. Increase sample_count or "
+                "Not enough waypoints for Twist. Increase sample_count or "
                 "decrease hand_interp_steps."
             )
-        base, remainder = divmod(motion_count, 3)
-        values = [base + (index < remainder) for index in range(3)]
-        return values[0], values[1], values[2]
+        base, remainder = divmod(motion_count, 4)
+        values = [base + (index < remainder) for index in range(4)]
+        return values[0], values[1], values[2], values[3]
 
     def _plan_pose_segment(
         self,
         target_pose: torch.Tensor,
         start_qpos: torch.Tensor,
         control_part: str,
-        request: ResolvedActionRequest[PressButtonGoal, PressButtonOptions],
+        request: ResolvedActionRequest[TwistGoal, TwistOptions],
         sample_count: int,
     ) -> tuple[torch.Tensor, torch.Tensor]:
         result = self.motion_generator.generate(
@@ -321,5 +364,35 @@ class PressButton(AtomicAction[PressButtonGoal, PressButtonOptions]):
         assert result.positions is not None
         return result.success, result.positions
 
+    def _twisted_grasp_poses(
+        self,
+        link_pose: torch.Tensor,
+        grasp_xpos: torch.Tensor,
+        twist_axis: torch.Tensor,
+        twist_angle: float,
+        waypoint_count: int,
+    ) -> torch.Tensor:
+        """Build Cartesian EEF keyframes that follow the target's twist arc."""
+        axis = twist_axis.to(device=self.device, dtype=torch.float32)
+        axis = axis / torch.linalg.vector_norm(axis)
+        angles = torch.linspace(
+            twist_angle / waypoint_count,
+            twist_angle,
+            waypoint_count,
+            dtype=torch.float32,
+            device=self.device,
+        )
+        rotations = (
+            torch.eye(4, dtype=torch.float32, device=self.device)
+            .reshape(1, 4, 4)
+            .repeat(waypoint_count, 1, 1)
+        )
+        rotations[:, :3, :3] = axis_angle_to_rotation_matrix(angles[:, None] * axis)
+        link_to_eef = torch.bmm(pose_inv(link_pose), grasp_xpos)
+        return torch.matmul(
+            torch.matmul(link_pose[:, None], rotations[None]),
+            link_to_eef[:, None],
+        )
 
-__all__ = ["PressButton", "PressButtonGoal", "PressButtonOptions"]
+
+__all__ = ["Twist", "TwistGoal", "TwistOptions"]
